+++ src/index.js
@@ -1,0 +1,7 @@
+// @flow
+
+import update from './update';
+import wrapCursor from './wrapCursor';
+
+module.exports = update;
+(module.exports: any).wrapCursor = wrapCursor;